remove: Delivery 관련 객체들 옮기고 State도 분리
