+++ service/order/order_server/src/main/java/com/sparta/order/domain/model/Order.java
@@ -1,20 +1,16 @@
 package com.sparta.order.domain.model;
 
 import com.sparta.commons.domain.jpa.BaseEntity;
-import com.sparta.order.domain.model.State.OrderState;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
 import jakarta.persistence.OneToMany;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -58,9 +54,8 @@ public class Order extends BaseEntity {
 
   private int totalQuantity = 0;
 
-  @OneToOne(mappedBy = "order", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-  @JoinColumn(name = "delivery_id")
-  private Delivery delivery;
+  @Column(nullable = false)
+  private UUID deliveryId;
 
   private boolean isDelete = false;
 
@@ -81,8 +76,8 @@ public class Order extends BaseEntity {
     this.orderState = state;
   }
 
-  public void setDelivery(Delivery delivery) {
-    this.delivery = delivery;
+  public void setDeliveryId(UUID deliveryId) {
+    this.deliveryId = deliveryId;
   }
 
   private Money calculateTotalAmount(List<OrderDetail> orderDetailList) {
